fix(meeting-bot): keep realtime config API stable
